Add message queues for updates to gain cache entries
Threads that do not own nodes should not update the gain cache entries
of these nodes. The message queues allow threads to perform gain cache
updates for their nodes that are neighbors of moved vertices from other
threads. There is currently no mutex machanic tied to the message queues
as it is not essential to catch all messages immediately. The queue is
FIFO, so no messages are lost and therefore consumed at some point.
There are also no write race conditions to queue entries.