Use barrier to handle message queues
The barrier is acting like two barriers. It is meant to be used to sync
threads at the beginning of a block and then again at the end of this
block. Also supports decrementing the size if threads finish and
therefore will not reenter the block, which otherwise would produce a
livelock.